feat: background image 크기 조정을 위한 컴포넌트 재배치
