+++ src/components/MessageCardList/CardListBackground/CardListBackground.module.css
@@ -1,5 +1,6 @@
 .CardListBackground {
   padding: 0 2.4rem;
+  height: 100vh;
 }
 
 .blue {
