+++ src/components/MessageCardList/MessageCardList.module.css
@@ -1,3 +1,6 @@
+.CardListPosition {
+}
+
 .CardListPadding {
   max-width: 120rem;
   height: 11rem;
