+++ src/pages/RecipientPage.module.css
@@ -1,0 +1,6 @@
+.FixPosition {
+  position: sticky;
+  top: 0;
+  background-color: white;
+  z-index: 1;
+}
